attempted implementation of web stuff
it doesnt work but it will soon....

the current error is that it says that selenium doesnt work on the jar file that i imported.

also, the "Sensitive" class contains my login details so i wont be publishing them; in an end user scenario, just add your username and password into those fields.